helpers: new module for common helper functions
